+++ marshaler.go
@@ -1,0 +1,7 @@
+package marshalkit
+
+// Marshaler is a marshaler that can marshal types and values.
+type Marshaler interface {
+	TypeMarshaler
+	ValueMarshaler
+}